refactor: 뱃지 계산 로직 custom hook으로 분리
